+++ SerialXL320.cpp
@@ -3,7 +3,7 @@
 
 /* ******************************** PUBLIC METHODS ************************************** */
 
-SerialXL320::SerialXL320(const DnxHAL::Port_t& port_in, int baud_in, int return_level_in /*=1*/) :
+SerialXL320::SerialXL320(const DnxHAL::Port_t& port_in, long int baud_in, int return_level_in /*=1*/) :
     DnxHAL(port_in, baud_in, return_level_in){
     if(debug_) fprintf(fp_debug_, "XL320_SERIAL: Object attached to serial at baud rate %ld and bit period of %f us\n\r", baud_in, bit_period_);
 }
